+++ api/test/SubmitCode.test.ts
@@ -1,11 +1,13 @@
 import { readFileSync } from 'fs';
 import { join } from 'path';
+
 import { getAccount, sendTransaction, sleep } from './utilsFunctions';
-import { GearApi } from '../lib';
+import { GearApi } from '../src';
 import { GEAR_EXAMPLES_WASM_DIR } from './config';
 
 const api = new GearApi();
 const accounts = {};
+const code = readFileSync(join(GEAR_EXAMPLES_WASM_DIR, 'demo_sum.opt.wasm'));
 
 beforeAll(async () => {
   await api.isReady;
@@ -19,13 +21,16 @@ afterAll(async () => {
 
 describe('Submit code', () => {
   test('demo_sum', async () => {
-    const code = readFileSync(join(GEAR_EXAMPLES_WASM_DIR, 'demo_sum.opt.wasm'));
-    const { codeHash } = api.code.submit(code);
+    const { codeHash } = await api.code.submit(code);
     expect(codeHash).toBeDefined();
-
     const transactionData = await sendTransaction(api.code, accounts['alice'], 'CodeChanged');
     expect(transactionData.id).toBe(codeHash);
     expect(transactionData.change).toHaveProperty('Active');
     expect(transactionData.change.Active).toHaveProperty('expiration');
+  });
+
+  test('Throw error when code exists', async () => {
+    await expect(api.code.submit(code),
+    ).rejects.toThrow('Code already exists');
   });
 });